+++ src/components/ids-data-grid/ids-data-grid-formatters.ts
@@ -1,7 +1,17 @@
 import '../ids-hyperlink/ids-hyperlink';
 import '../ids-button/ids-button';
-import '../ids-badge/ids-badge';
 import '../ids-alert/ids-alert';
+import '../ids-badge/ids-badge';
+import '../ids-card/ids-card';
+import '../ids-color/ids-color';
+import '../ids-counts/ids-counts';
+import '../ids-icon/ids-icon';
+import '../ids-image/ids-image';
+import '../ids-progress-bar/ids-progress-bar';
+import '../ids-rating/ids-rating';
+import '../ids-slider/ids-slider';
+import '../ids-step-chart/ids-step-chart';
+import '../ids-tag/ids-tag';
 
 import { stringToBool } from '../../utils/ids-string-utils/ids-string-utils';
 import { escapeHTML } from '../../utils/ids-xss-utils/ids-xss-utils';
@@ -47,6 +57,18 @@ export default class IdsDataGridFormatters {
     return typeof color === 'function' ? color(row, value, col, item) : color;
   }
 
+  #readonly(row: number, value: any, col: IdsDataGridColumn, item: Record<string, any>): boolean | undefined {
+    const readonly = col.readonly;
+
+    return typeof readonly === 'function' ? readonly(row, value, col, item) : readonly;
+  }
+
+  #size(row: number, value: any, col: IdsDataGridColumn, item: Record<string, any>): string | undefined {
+    const size = col.size;
+
+    return typeof size === 'function' ? size(row, value, col, item) : size;
+  }
+
   /** Displays a Text String */
   text(rowData: Record<string, unknown>, columnData: IdsDataGridColumn): string {
     return `<span class="text-ellipsis">${this.#extractValue(rowData, columnData.field)}</span>`;
@@ -66,14 +88,16 @@ export default class IdsDataGridFormatters {
   date(rowData: Record<string, unknown>, columnData: IdsDataGridColumn, index: number, api: IdsDataGrid): string {
     let value: any = this.#extractValue(rowData, columnData.field);
     value = api.localeAPI?.formatDate(value, columnData.formatOptions) ?? value.toString();
-    return `<span class="text-ellipsis">${value}</span>`;
+    const icon = columnData.editor?.type === 'datepicker' ? '<ids-icon icon="schedule" class="editor-cell-icon"></ids-icon>' : '';
+    return `<span class="text-ellipsis">${value}</span>${icon}`;
   }
 
   /** Formats date data as a time string in the desired format */
   time(rowData: Record<string, unknown>, columnData: IdsDataGridColumn, index: number, api: IdsDataGrid): string {
     let value: any = this.#extractValue(rowData, columnData.field);
     value = api.localeAPI?.formatDate(value, columnData.formatOptions || { timeStyle: 'short' }) ?? value.toString();
-    return `<span class="text-ellipsis">${value}</span>`;
+    const icon = columnData.editor?.type === 'timepicker' ? '<ids-icon icon="clock" class="editor-cell-icon"></ids-icon>' : '';
+    return `<span class="text-ellipsis">${value}</span>${icon}`;
   }
 
   /** Formats number data as a decimal string in the specific localeAPI */
@@ -160,7 +184,7 @@ export default class IdsDataGridFormatters {
 
   /** Shows an Tree */
   tree(rowData: Record<string, unknown>, columnData: IdsDataGridColumn): string {
-    const value: any = this.#extractValue(rowData, columnData.field);
+    const value = this.#extractValue(rowData, columnData.field);
     const button = rowData?.children ? `<ids-button tabindex="-1" class="expand-button">
       <ids-icon icon="plusminus-folder-${rowData.rowExpanded === false ? 'closed' : 'open'}"></ids-icon>
     </ids-button>` : '&nbsp;';
@@ -176,6 +200,7 @@ export default class IdsDataGridFormatters {
     return `<span class="ids-data-grid-tree-container">${button}<span class="text-ellipsis">${value}</span></span>`;
   }
 
+  /** Shows a dropdown list */
   dropdown(rowData: Record<string, unknown>, columnData: IdsDataGridColumn): string {
     const field = columnData.field ?? '';
     const options = <any[]>columnData.editor?.editorSettings?.options || [];
@@ -191,8 +216,251 @@ export default class IdsDataGridFormatters {
       </span>
       <ids-icon
         icon="dropdown"
-        class="dropdown-cell-icon">
+        class="editor-cell-icon">
       </ids-icon>
     `;
+  }
+
+  /* Shows ids-alert, and the field value will appear in a tooltip. An `icon` option can be provided as an override. */
+  alert(rowData: Record<string, unknown>, columnData: IdsDataGridColumn): string {
+    const value: any = this.#extractValue(rowData, columnData.field);
+    if (!value && !columnData.icon) return '';
+
+    const icon = columnData.icon ?? 'alert';
+    const tooltip = value ? `tooltip="${value}"` : '';
+
+    return `
+      <span class="text-ellipsis">
+        <ids-alert icon="${icon}" ${tooltip}></ids-alert>
+      </span>
+    `;
+  }
+
+  /* Shows ids-color. If a `color` option is provided, the field's value will appear in a tooltip. */
+  color(rowData: Record<string, unknown>, columnData: IdsDataGridColumn, index: number): string {
+    const value: any = this.#extractValue(rowData, columnData.field);
+    if (!columnData.color && !value) return '<span class="text-ellipsis"><ids-color></ids-color></span>';
+
+    const color = this.#color(index, value, columnData, rowData);
+
+    const hex = color || value || '#C2A1F1';
+    const tooltip = !color && value ? `tooltip="${value}"` : '';
+
+    return `
+      <span class="text-ellipsis">
+        <ids-color hex="${hex}" ${tooltip}></ids-color>
+      </span>
+    `;
+  }
+
+  /* Shows the field value as an ids-icon. An `icon` and `size` option can be provided as overrides. */
+  icon(rowData: Record<string, unknown>, columnData: IdsDataGridColumn, index: number): string {
+    const value: any = this.#extractValue(rowData, columnData.field);
+    if (!value && !columnData.icon) return '';
+
+    const color = this.#color(index, value, columnData, rowData) || '';
+    const badgeColor = color ? `badge-color="${color}" badge-position="top-left"` : '';
+
+    const sizes = ['small', 'medium', 'large', 'xl', 'xxl'];
+    let size = this.#size(index, value, columnData, rowData) || '';
+    size = sizes.includes(size) ? size : 'large';
+
+    const icon = String(columnData.icon || value).replace('icon-', '');
+    const text = (columnData.icon && typeof value === typeof '') ? value : '';
+
+    return `
+      <span class="text-ellipsis">
+        <ids-icon icon="${icon}" size="${size}" ${badgeColor}></ids-icon>${text}
+      </span>
+    `;
+  }
+
+  /* Shows the field value as a `star-filled` or `star-outlined`. A `size` option can be provided as an override. */
+  favorite(rowData: Record<string, unknown>, columnData: IdsDataGridColumn, index: number): string {
+    const value: any = this.#extractValue(rowData, columnData.field);
+
+    return this.rating(
+      {
+        ...rowData,
+        [String(columnData.field)]: value ? 1 : 0,
+      },
+      {
+        ...columnData,
+        max: 1,
+        min: 0,
+      },
+      index,
+    );
+  }
+
+  /* Shows the field value as an `ids-tag`. A `color` option can be provdied as an override. */
+  tag(rowData: Record<string, unknown>, columnData: IdsDataGridColumn, index: number): string {
+    const value: any = this.#extractValue(rowData, columnData.field);
+    if (!value) return '';
+    const color = this.#color(index, value, columnData, rowData);
+
+    return `
+      <span class="text-ellipsis">
+        <ids-tag color="${color || ''}">${value}</ids-tag>
+      </span>
+    `;
+  }
+
+  /*
+    Shows the field value as an `ids-progress`. A `text` option can be provided to customize the label.
+    A `color` and `max` option can be provided as overrides.
+  */
+  progress(rowData: Record<string, unknown>, columnData: IdsDataGridColumn): string {
+    const value: any = parseFloat(this.#extractValue(rowData, columnData.field));
+    // const color = this.#color(index, value, columnData, rowData);
+
+    const val = Number.isNaN(value) ? 0 : value;
+    let max = columnData.max ?? 10;
+    if (!columnData.max && value > 1) {
+      const digitCount = String(Math.floor(value)).length;
+      max = parseInt(`1`.padEnd(digitCount + 1, '0'));
+    }
+
+    // TODO: Fix label and label-audible attribute
+    const label = `label="${columnData?.text || `${val} of ${max}`}" label-audible`;
+
+    return `
+      <ids-progress-bar ${label} max="${max}" value="${val}" readonly>
+      </ids-progress-bar>
+    `;
+  }
+
+  /*
+    Shows the field value as an `ids-rating`. A `text` option can be provided to customize the label.
+    A `color` and `max` option can be provided as overrides.
+  */
+  rating(rowData: Record<string, unknown>, columnData: IdsDataGridColumn, index: number): string {
+    const value: any = parseFloat(this.#extractValue(rowData, columnData.field));
+    const color = this.#color(index, value, columnData, rowData);
+
+    const max = columnData.max ?? 5;
+    let val = Number.isNaN(value) ? 0 : value;
+    if ((!columnData.max && val > 1) || val > max) {
+      const digitCount = String(Math.floor(val)).length;
+      const divisor = parseInt(`1`.padEnd(digitCount + 1, '0'));
+
+      if (val > 10) {
+        val = (val / divisor) * max;
+      }
+    }
+
+    const sizes = ['small', 'medium', 'large', 'xl', 'xxl'];
+    let size = this.#size(index, value, columnData, rowData) || '';
+    size = sizes.includes(size) ? size : 'large';
+
+    const label = columnData.text ?? `${val} of ${max} stars`;
+    const readonly = this.#readonly(index, value, columnData, rowData) ?? true;
+
+    return `
+      <span class="text-ellipsis">
+        <ids-rating
+          label="${label}"
+          size="${size}"
+          stars="${max}"
+          value="${val}"
+          ${color ? `color="${color}"` : ''}
+          ${readonly ? 'readonly' : ''}
+        >
+        </ids-rating>
+      </span>
+    `;
+  }
+
+  /*
+    Shows the field value as an `ids-slider`. A `text` option can be provided to customize the label.
+    A `color`, `max`, `min` and `type` option can be provided as overrides.
+  */
+  slider(rowData: Record<string, unknown>, columnData: IdsDataGridColumn, index: number): string {
+    const value: any = parseFloat(this.#extractValue(rowData, columnData.field));
+    const color = this.#color(index, value, columnData, rowData);
+
+    const min = columnData.min ?? 0;
+    const max = columnData.max ?? 100;
+    let val = Number.isNaN(value) ? 0 : value;
+    if ((!columnData.max && val > 1) || val > max) {
+      const digitCount = String(Math.floor(val)).length;
+      const divisor = parseInt(`1`.padEnd(digitCount + 1, '0'));
+
+      if (val > 10) {
+        val = (val / divisor) * max;
+      }
+    }
+
+    const type = columnData.type ?? 'single';
+    const label = columnData.text ?? `${val} of ${max} stars`;
+    const readonly = this.#readonly(index, value, columnData, rowData) ?? true;
+
+    return `
+      <ids-slider
+        type="${type}"
+        label="${label}"
+        min="${Math.max(min, 0)}"
+        max="${max}"
+        value="${Math.min(val, max)}"
+        show-tooltip
+        ${color ? `color="${color}"` : ''}
+        ${readonly ? 'readonly' : ''}
+      >
+      </ids-slider>
+    `;
+  }
+
+  /*
+    Shows the field value as an `ids-step-chart`. A `text` option can be provided to customize the label.
+    A `color` and `max` option can be provided as overrides.
+  */
+  stepChart(rowData: Record<string, unknown>, columnData: IdsDataGridColumn, index: number): string {
+    const value: any = parseFloat(this.#extractValue(rowData, columnData.field));
+    const color = this.#color(index, value, columnData, rowData);
+
+    let max = columnData.max ?? 5;
+    if (!columnData.max && value > 1) {
+      const digitCount = String(Math.floor(value)).length;
+      max = parseInt(`1`.padEnd(digitCount + 1, '0'));
+    }
+
+    let val = Number.isNaN(value) ? 0 : value;
+    if (max > 10) {
+      val = (val / max) * 10;
+      max = 10;
+    }
+
+    // const label = columnData.text ?? `${val} of ${max} steps completed`;
+    const completedSteps = Math.floor(val);
+    const stepsInProgress = Math.ceil(val);
+    const showStepsInProgress = completedSteps !== stepsInProgress;
+
+    return `
+      <ids-step-chart
+        step-number="${Math.min(max, 10)}"
+        steps-in-progress="${showStepsInProgress ? stepsInProgress : 0}"
+        value="${Math.floor(val)}"
+        progress-color="ruby02"
+        ${color ? `color="${color}"` : ''}
+      >
+      </ids-step-chart>
+    `;
+  }
+
+  /* Shows the field value as an `ids-image`. A `text` option can be provided to the `alt` and `title` attributes. */
+  image(rowData: Record<string, unknown>, columnData: IdsDataGridColumn): string {
+    const value: any = this.#extractValue(rowData, columnData.field);
+    if (!value) return '';
+
+    const metadata = columnData.text ? `alt="${columnData.text}" title="${columnData.text}"` : '';
+
+    return `<ids-image src="${value}" ${metadata}></ids-image>`;
+  }
+
+  card(rowData: Record<string, unknown>, columnData: IdsDataGridColumn): string {
+    const value: any = this.#extractValue(rowData, columnData.field);
+    if (!value) return '';
+
+    return `<ids-card>${value}</ids-card>`;
   }
 }