+++ src/components/ids-date-picker/ids-date-picker.ts
@@ -154,7 +154,7 @@ class IdsDatePicker extends Base {
    * List of available color variants for this component
    * @returns {Array<string>}
    */
-  colorVariants = ['alternate-formatter'];
+  colorVariants = ['alternate-formatter', 'borderless', 'in-cell'];
 
   /**
    * Push color variant to the trigger-field element
@@ -307,6 +307,7 @@ class IdsDatePicker extends Base {
       this.#picker.onOutsideClick = (e: Event) => {
         if (this.#picker) {
           if (!e.composedPath()?.includes(this.#picker)) {
+            this.triggerEvent('outsideclick.datepicker', this);
             this.#togglePopup(false);
           }
         }